+++ src/Sentry.Profiling/SamplingTransactionProfiler.cs
@@ -96,11 +96,14 @@ internal class SamplingTransactionProfiler : ITransactionProfiler
     }
 
     /// <inheritdoc />
-    public async Task<ProfileInfo> CollectAsync(Transaction transaction)
+    public Protocol.Envelopes.ISerializable? Collect(Transaction transaction)
+        => Protocol.Envelopes.AsyncJsonSerializable.CreateFrom(CollectAsync(transaction));
+
+    internal async Task<ProfileInfo> CollectAsync(Transaction transaction)
     {
         if (!_stopped)
         {
-            throw new InvalidOperationException("Profiler.CollectAsync() called before Finish()");
+            throw new InvalidOperationException("Profiler.Collect() called before Finish()");
         }
 
         // Wait for the last sample (<= _endTimeMs), or at most 1 second. The timeout shouldn't happen because